Modify pipe::push() to reflect its documentation.
pipe::push() claimed that "If this call is unsuccessful, packet_p will still point to the packet after the call." but that's not what its implementation did. It would just discard the packet. Its signature was changed to take the unique_ptr<> by reference instead so that the caller will still have the data in the packet in case the function returns false.